requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.3
django-ckeditor==6.0.0
django-js-asset==1.2.2
ez-setup==0.9
gunicorn==20.1.0
pkg-resources==0.0.0
psycopg2==2.7.7
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.2.0
